Adding 'WIFI DART' to all the device tree model strings
Variscite's WIFI and Bluetooth startup scripts check to see if you are
running the version of the SoM with WIFI and/or BT by checking the model
description in the device tree and see if it mentions "WIFI" and "DART".
This seems like a terrible way to check the hardware, but they closed
out the bug that I filed as "won't fix" on their end, so I guess we have
to add the right keywords to the description to get the hardware
working.